ENH:  expand_dims
